float

+++ assets/js/custom.js
@@ -39,12 +39,12 @@ myHash['Saturday'];
 myHash['Sunday'];
 
 function calcHour(day) {
-	var timeOut = parseInt(document.getElementById(day.concat("TimeOut")).value);
-	var timeIn = parseInt(document.getElementById(day.concat("TimeIn")).value);
-	var lunchOut = parseInt(document.getElementById(day.concat("LunchOut")).value);
-	var lunchIn = parseInt(document.getElementById(day.concat("LunchIn")).value);
-	var workTime = parseInt(timeOut - timeIn);
-	var lunchTime = parseInt(lunchOut - lunchIn);
+	var timeOut = parseFloat(document.getElementById(day.concat("TimeOut")).value);
+	var timeIn = parseFloat(document.getElementById(day.concat("TimeIn")).value);
+	var lunchOut = parseFloat(document.getElementById(day.concat("LunchOut")).value);
+	var lunchIn = parseFloat(document.getElementById(day.concat("LunchIn")).value);
+	var workTime = parseFloat(timeOut - timeIn);
+	var lunchTime = parseFloat(lunchOut - lunchIn);
 
 	if (workTime < 0 || lunchTime < 0 || workTime-lunchTime < 0 || lunchOut > timeOut) {
 		myHash[day] = "Error";
